+++ finagle-core/src/main/scala/com/twitter/finagle/client/LatencyCompensation.scala
@@ -7,11 +7,11 @@ import com.twitter.util.Duration
 import java.util.concurrent.atomic.AtomicReference
 
 /**
- * Latency compensation enables the modification of connection and
- * request timeouts on a per-endpoint basis.  For instance, if a
- * client has both network-local and trans-continental endpoints, a
- * reasonable latency compensator might add the speed-of-light penalty
- * when communicating with distant endpoints.
+ * Latency compensation enables the modification of connection, request, and
+ * session acquisition timeouts on a per-endpoint basis.  For instance, if a client
+ * has both network-local and trans-continental endpoints, a reasonable latency
+ * compensator might add the speed-of-light penalty when communicating with distant
+ * endpoints.
  */
 object LatencyCompensation {